memset/memcpy using too-small type for index

+++ tos/platforms/bacon/bacon_mem.h
@@ -2,7 +2,7 @@
 #define BREAKFAST_MEM_H
 
 void*  memcpy(void *dest, const void *src, size_t n){
-  uint8_t i;
+  size_t i;
   uint8_t* d =(uint8_t*)dest;
   uint8_t* s =(uint8_t*)src;
   for (i =0; i< n; i++){
@@ -12,7 +12,7 @@ void*  memcpy(void *dest, const void *src, size_t n){
 }
 
 void* memset(void *dest, int c, size_t n){
-  uint8_t i;
+  size_t i;
   uint8_t* d = (uint8_t*)dest;
   for(i=0; i < n; i++){
     d[i] = c;
